move python files
