+++ src-python/src/map_objects.py
